requirements.txt
fastapi
uvicorn
authlib
pyjwt
flake8